Gather computer information and FTP that info, establish reverse shell - Language Override
This is a DuckyScript for Windows originally downloaded from https://ducktoolkit.com, which overrides any local keyboard settings and temporarily switches the keyboard to English - DuckyScript is only compatible with English keyboards! Info gathered includes the following:
1) Computer information
2) USB information
3) User information
4) Shared drive information
5) Installed program information
6) Installed updates
7) Network information
8) Wireless profile
9) SAM file

Then it will upload all that information to the FTP server of your choice (line 217 or thereabouts), then pop a reverse shell. 

The reverse shell is still a work in progress and could only get it to work intermittently, but this project is not done yet.

The license for this script is 'I AM NOT RESPONSIBLE' license v 1.0